refactor(docs): move refresh authorization types
