+++ Web/ForgeTrust.AppSurface.Docs/Controllers/DocsController.cs
@@ -1841,28 +1841,3 @@ public class DocsController : Controller
         public IFileProvider ContentRootFileProvider { get; set; } = new NullFileProvider();
     }
 }
-
-internal readonly record struct SearchIndexRefreshAuthorizationResult(
-    bool IsAllowed,
-    SearchIndexRefreshAuthorizationFailure? Reason)
-{
-    public static SearchIndexRefreshAuthorizationResult Allowed()
-    {
-        return new SearchIndexRefreshAuthorizationResult(true, null);
-    }
-
-    public static SearchIndexRefreshAuthorizationResult Denied(SearchIndexRefreshAuthorizationFailure reason)
-    {
-        return new SearchIndexRefreshAuthorizationResult(false, reason);
-    }
-}
-
-internal enum SearchIndexRefreshAuthorizationFailure
-{
-    MissingPolicyOption,
-    MissingPolicyProvider,
-    MissingAuthorizationService,
-    PolicyNotFound,
-    Unauthenticated,
-    AuthorizationFailed
-}
